+++ src/components/Proposal/Update/EmptyProposalUpdate.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import TokenList from 'decentraland-gatsby/dist/utils/dom/TokenList'
 
-import QuestionCircleIcon from '../Icon/QuestionCircle'
+import QuestionCircleIcon from '../../Icon/QuestionCircle'
 
 import './ProposalUpdate.css'
