+++ src/components/Proposal/Update/ProposalUpdates.tsx
@@ -4,10 +4,10 @@ import Paragraph from 'decentraland-gatsby/dist/components/Text/Paragraph'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import { Header } from 'decentraland-ui/dist/components/Header/Header'
 
-import { ProposalAttributes } from '../../entities/Proposal/types'
-import { UpdateAttributes } from '../../entities/Updates/types'
-import Megaphone from '../Icon/Megaphone'
-import Divider from '../Section/Divider'
+import { ProposalAttributes } from '../../../entities/Proposal/types'
+import { UpdateAttributes } from '../../../entities/Updates/types'
+import Megaphone from '../../Icon/Megaphone'
+import Divider from '../../Section/Divider'
 
 import ProposalUpdate from './ProposalUpdate'
 import './ProposalUpdates.css'
