+++ src/pages/proposal.tsx
@@ -33,7 +33,7 @@ import { VotesList } from '../components/Modal/Votes/VotesList'
 import ProposalComments from '../components/Proposal/ProposalComments'
 import ProposalFooterPoi from '../components/Proposal/ProposalFooterPoi'
 import ProposalHeaderPoi from '../components/Proposal/ProposalHeaderPoi'
-import ProposalUpdates from '../components/Proposal/ProposalUpdates'
+import ProposalUpdates from '../components/Proposal/Update/ProposalUpdates'
 import ForumButton from '../components/Section/ForumButton'
 import ProposalDetailSection from '../components/Section/ProposalDetailSection'
 import ProposalResultSection from '../components/Section/ProposalResultSection'
